fixes for PR
